Some improvements (#44)
* Support for running in non-coroutine

* Update test

* Fix multiple definition of `startTime'

* Add test

* test strict

* Fix type

* Fix segmentation fault when not connected

* Fix test

* Add swoole 5.0 test

* Add watch test

* Add ZOO_LOG_LEVEL constants and set default debug level

* FIx Segmentation fault (core dumped)

* FIx Segmentation fault (core dumped)

+++ src/php/api.stub.php
@@ -7,7 +7,7 @@ namespace Swoole {
 
     class ZooKeeper
     {
-        public function __construct(string $host, double $timeout)
+        public function __construct(string $host, float $timeout)
         {
         }
